Revert "Create: `utils.c`/`.h` for easier maintainability"
This reverts commit 87afcc6d370a8e32b71503ada3506df9b5f923a8.
